Added a splash page
Bit trickier than adding a splash page to the pomodoro project as I can only hide the splash page when all ajax requests were
complete. Luckily there is an ajaxStop method that waits for the completion of the last outstanding ajax call.

Also replaced the background with a solid colour.